+++ lib/http/apps/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  space: require('./space'),
+  dao: require('./dao'),
   wallet: require('./wallet'),
   generic: require('./generic')
 }
