+++ lib/http/index.js
@@ -17,7 +17,7 @@ const app = express();
 const router = express.Router();
 
 const walletApp = require('./apps/wallet');
-const spaceApp = require('./apps/space');
+const daoApp = require('./apps/dao');
 const genericApp = require('./apps/generic');
 
 let ethProvider;
@@ -27,7 +27,6 @@ if (process.env.ETH_PROVIDER_URL) {
   ethProvider = new ethers.getDefaultProvider('rinkeby'); // TODO: use correct network
 }
 
-ethProvider.getSigner();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
@@ -38,9 +37,9 @@ app.use(
   walletApp
 );
 app.use(
-  '/space',
+  '/dao',
   middlewares.loadCoinsenceKit({ password, ethProvider }),
-  spaceApp
+  daoApp
 );
 app.use(
   '/',
